# api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: 'Avaliação Prática Lemobs'
  version: '1.0.0'
  title: Desafio API Lemobs
basePath: /v1/
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
paths:
...
  /endereco:
    x-swagger-router-controller: endereco
    post:
      summary: Cadastrar um novo endereço.
      tags:
        - Endereço
      operationId: cadastrar_endereco
      parameters:
        - in: body
          name: body
          description: Modelo de Endereço
          required: true
          schema:
            $ref: "#/definitions/Endereco"
      responses:
        '201':
          description: Criado (Created)
        '400':
          description: Requisição Inválida (Bad Request)
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      summary: Listar todos os endereços.
      tags:
        - Endereço
      operationId: listar_enderecos
      parameters:
        - in: query
          name: bairro
          description: Bairro para filtrar Endereço
          type: string
      responses:
        '200':
          description: Sucesso (OK)
          schema:
            $ref: "#/definitions/EnderecosResponse"
        '400':
          description: Requisição Inválida (Bad Request)
          schema:
            $ref: "#/definitions/ErrorResponse"
        '500':
          description: Erro Interno do Servidor (Internal Server Error)
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Endereco:
    type: object
    properties:
      rua:
        type: string
      numero:
        type: string
      complemento:
        type: string
      bairro:
        type: string
      aluno_id:
        type: integer
    required:
      - rua
      - bairro
      - aluno_id
...
  EnderecosResponse:
    required:
      - total
      - enderecos
    properties:
      total:
        type: integer
      enderecos:
        type: array
        items:
          type: object
          properties:
            endereco:
              type: string
            bairro:
              type: string
          required:
            - endereco
            - bairro
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string